+++ Apex.GameZone.Core/Models/ProductModel.cs
@@ -4,8 +4,8 @@ public class ProductModel : BaseModel
 {
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal Price { get; set; }
-    public decimal ActualPrice { get; set; }
+    public int Price { get; set; }
+    public int ActualPrice { get; set; }
     public int Quantity { get; set; }
     public int GameZoneId { get; set; }
 }
